+++ fractal/components/_preview-layouts/_recommendation-arc-chart-preview.jsx
@@ -7,8 +7,8 @@ const page = ({ rendered_content, path }) => {
             <link media="all" rel="stylesheet" href={`http://localhost:3050/static/css/core-configuration.css`}/>
             <link media="all" rel="stylesheet" href={`http://localhost:3050/static/css/charts/SVGArcChart/SVGArcChart.css`}/>
             <div className="base-preview" dangerouslySetInnerHTML={{__html: rendered_content}}/>
-            {/*<script src={`http://localhost:3050/static/vendors/snap.svg-min.js`}/>*/}
-            <script src={`http://localhost:3050/static/js/SVGArcChart.js`}/>
+            <script src={`http://localhost:3050/static/vendors/snap.svg-min.js`}/>
+            <script src={`http://localhost:3050/static/js/RecommendationChart.js`}/>
         </div>
     );
 };