+++ fractal/components/_preview-layouts/_recommendation-bar-chart-preview.jsx
@@ -5,10 +5,8 @@ const page = ({ rendered_content, path }) => {
     return (
         <div>
             <link media="all" rel="stylesheet" href={`http://localhost:3050/static/css/core-configuration.css`}/>
-            <link media="all" rel="stylesheet" href={`http://localhost:3050/static/css/charts/SVGArcChart/SVGArcChart.css`}/>
+            <link media="all" rel="stylesheet" href={`http://localhost:3050/static/css/charts/SVGBarChart/SVGBarChart.css`}/>
             <div className="base-preview" dangerouslySetInnerHTML={{__html: rendered_content}}/>
-            {/*<script src={`http://localhost:3050/static/vendors/snap.svg-min.js`}/>*/}
-            <script src={`http://localhost:3050/static/js/SVGArcChart.js`}/>
         </div>
     );
 };