+++ fractal/components/assessments/RecommendationChart/index.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+
+import RecommendationChart from './RecommendationChart.jsx';
+import config from './RecommendationChart.config';
+
+ReactDOM.render(
+    <RecommendationChart {...config.context} />,
+    document.getElementsByClassName('base-preview')[0]
+);