+++ fractal/components/assessments/RecommendationSummary/index.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+
+import RecommendationSummary from './RecommendationSummary.jsx';
+import config from './RecommendationSummary.config';
+
+ReactDOM.render(
+    <RecommendationSummary {...config.context} />,
+    document.getElementsByClassName('base-preview')[0]
+);