+++ fractal/components/charts/SVGArcChart/index.jsx
@@ -1,0 +1,17 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+
+import SVGArcChart from './SVGArcChart.jsx';
+import config from './SVGArcChart.config';
+
+ReactDOM.render(
+    <div>
+        <SVGArcChart {...config.context} />
+        {
+            config.variants.map((variant) => {
+                return <SVGArcChart key={variant.name} {...variant.context}/>
+            })
+        }
+    </div>,
+    document.getElementsByClassName('base-preview')[0]
+);