+++ fractal/components/loading-spinners/LoadingSpinner/LoadingSpinner.config.js
@@ -1,7 +1,7 @@
 module.exports = {
     "title": "Loading Spinner",
     "status": "ready",
-    "collated": true,
+    "collated": false,
     "preview": "@loading-spinner-preview",
     "context": {
         "text": "Please wait..."
@@ -9,6 +9,9 @@ module.exports = {
     "variants": [
         {"name": "loading-spinner-overlay", "context": {
             "overlay": true
+        }},
+        {"name": "loading-spinner-fullscreen", "context": {
+            "fullscreen": true
         }}
     ]
 };