+++ fractal/components/loading-spinners/LoadingSpinner/LoadingSpinner.jsx
@@ -2,11 +2,13 @@ import React from 'react';
 import classNames from 'classnames';
 import PropTypes from 'prop-types';
 
-const LoadingSpinner = ({ overlay = false, text }) => {
+const LoadingSpinner = ({ overlay = false, fullscreen = false, text }) => {
     const componentClass = 'c-loading';
     const spinnerClass = `${componentClass}__spinner`;
-    const componentClasses = classNames(componentClass, {['is-overlay']: overlay});
-    const spinnerClasses = classNames(spinnerClass, {['is-overlay']: overlay});
+    const componentClasses = classNames(componentClass,
+        {['is-overlay']: overlay || fullscreen},
+        {['is-fullscreen']: fullscreen}
+    );
     const componentProps = { className: componentClasses };
 
     return (
@@ -15,13 +17,14 @@ const LoadingSpinner = ({ overlay = false, text }) => {
                 text &&
                 <div className={`${componentClass}__text`}>{text}</div>
             }
-            <div className={spinnerClasses}/>
+            <div className={spinnerClass}/>
         </div>
     );
 };
 
 LoadingSpinner.propTypes = {
     overlay: PropTypes.bool,
+    fullscreen: PropTypes.bool,
     text: PropTypes.string
 };