+++ gulp/build-configs/webpack.config.style-guide.js
@@ -5,7 +5,10 @@ module.exports = {
     target: 'web',
     entry: {
         ToolTip: './fractal/components/tooltips/ToolTip/index.jsx',
-        StickyBar: './fractal/components/stickies/StickyBar/index.jsx'
+        StickyBar: './fractal/components/stickies/StickyBar/index.jsx',
+        SVGArcChart: './fractal/components/charts/SVGArcChart/index.jsx',
+        RecommendationChart: './fractal/components/assessments/RecommendationChart/index.jsx',
+        RecommendationSummaryGroup: './fractal/components/assessments/RecommendationSummaryGroup/index.jsx'
     },
     output: {
         path: path.resolve(__dirname, '../../_site/js'),